middleware/requirements.txt
flask==3.0.0
google-cloud-bigquery==3.17.0
google-cloud-texttospeech==2.16.0
google-cloud-speech==2.26.0
google-generativeai==0.5.0
google-cloud-aiplatform==1.49.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0